Bubbled up `makeSplit`, `isSplittable` into `HoodieFileInputFormatBase` (those are not file-format specific)
